rag-fraud-detection-system/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pytest
faiss-cpu
sentence-transformers
transformers
torch
gradio
requests